feat(wasm): make the whole stack compile for wasm32-unknown-unknown
arboard has no platform module outside unix/windows/macos, so it never compiled
for wasm — and it was an unconditional dependency, which blocked the entire
stack rather than just clipboard. Gate it to native and split clipboard.rs along
the crate's existing `imp` idiom. The web handle is inert, which is the case the
API was already built to express: new() never fails, the operations are always
safe to call, and is_available() is there to explain an inert Copy button, so no
caller needs a cfg. navigator.clipboard is async and gesture-gated, so wiring it
up is a design question, not a missing function body.

Drop image's default features: they pull rayon (an OS thread pool) and avif ->
ravif/rav1e, none of which the Image widgets need.

Give matcha-ecs and matcha-ecs-widgets the workspace's build.rs + cfg_aliases
idiom, and on wasm enable wgpu/fragile-send-sync-non-atomic-wasm (bevy_ecs
requires Send + Sync of every Component and Resource, and has no non-Send
component storage to opt out into) and bevy_platform/web — without the latter,
bevy_tasks takes its std branch and never polls a task again after its first
Pending, and bevy_platform::time::Instant resolves to std::time::Instant, which
panics on wasm.

Two things wgpu's feature cannot reach, because they are not wgpu types: winit's
web window is Rc-based, and the wake closure in ModelHandle captures winit's
event-loop proxy. Both end up inside bevy components/resources, so they get
unsafe Send/Sync impls on web — the same bargain wgpu itself makes, under the
same condition. A compile_error guard in each crate turns "single-threaded wasm"
from a silent assumption into a build failure if `atomics` is ever enabled.

Note this compiles but does not yet run: ThreadDriver still calls
std::thread::spawn and parking_lot::Condvar::wait, both of which panic on wasm.
That is the next commit.
