+++ matcha-ecs-widgets/Cargo.toml
@@ -45,5 +45,16 @@ web-time = { workspace = true }
 fxhash = { workspace = true }
 parking_lot = { workspace = true }
 
+# NOTE: the `cfg` expression below must stay in sync with the `web` alias
+# defined in build.rs (`all(target_arch = "wasm32", target_os = "unknown")`).
+[target.'cfg(all(target_arch = "wasm32", target_os = "unknown"))'.dependencies]
+# This crate also holds wgpu handles in ECS resources (FontCtx's stencil cache,
+# ImageCtx). matcha-ecs enables the same feature and Cargo unifies them, but
+# declaring it here too keeps the requirement honest at the crate that has it.
+wgpu = { workspace = true, features = ["fragile-send-sync-non-atomic-wasm"] }
+
+[build-dependencies]
+cfg_aliases = { workspace = true }
+
 [lints]
 workspace = true
