+++ matcha-window/Cargo.toml
@@ -34,10 +34,15 @@ thiserror = { workspace = true }
 log = { workspace = true }
 enum-map = "2.7.3"
 web-time = { workspace = true }
-arboard = { version = "3.6", default-features = false }
 
 # NOTE: the `cfg` expressions below must stay in sync with the `web` alias
 # defined in build.rs (`all(target_arch = "wasm32", target_os = "unknown")`).
+
+# arboard has no platform module outside unix/windows/macos, so it does not
+# compile for wasm at all. See src/clipboard/web.rs for the replacement.
+[target.'cfg(not(all(target_arch = "wasm32", target_os = "unknown")))'.dependencies]
+arboard = { version = "3.6", default-features = false }
+
 [target.'cfg(all(target_arch = "wasm32", target_os = "unknown"))'.dependencies]
 wasm-bindgen-futures = { workspace = true }
 wasm-bindgen = { workspace = true }
